+++ new-dti-website-redesign/src/app/design-system/Sidebar.tsx
@@ -9,7 +9,7 @@ export default function Sidebar() {
   const pathname = usePathname();
 
   return (
-    <aside className="h-full border-l border-border-1 bg-background-1 fixed w-[260px] z-10">
+    <aside className="h-full border-l border-border-1 bg-background-1 fixed w-[260px] z-10 overflow-scroll">
       <nav className="flex flex-col">
         <div className="p-4 border-b-1 border-border-1">
           <h5>IDOL Design System</h5>
