+++ new-dti-website-redesign/src/app/design-system/components/card/page.tsx
@@ -5,6 +5,8 @@ import TestimonialCard from '../../../../components/TestimonialCard';
 import FeatureCard from '../../../../components/FeatureCard';
 import Marquee from '../../../../components/Marquee';
 import RocketIcon from '../icon/RocketIcon';
+import TimelineCard, { RecruitmentEvent } from '@/components/TimelineCard';
+import { MemberCard, MemberDetailsCard } from '@/components/TeamCard';
 
 const features = [
   {
@@ -41,6 +43,38 @@ const testimonials = [
   }
 ];
 
+const mockEvent: RecruitmentEvent = {
+  title: 'Applications open!',
+  description:
+    "We're welcoming any and all students who are looking to make a difference through tech. Applicants are not considered on a rolling basis.",
+  location: 'Application Link',
+  type: 'application',
+  link: 'https://docs.google.com/forms/d/e/1FAIpQLSc_Tp8BAM0ad-x5GmCRN_YgPOUwRjQynbGLuqQ28932zfwXrg/viewform?usp=sf_link',
+  freshmen: { date: 'August 21', isTentative: false, time: '5-6PM' },
+  upperclassmen: { date: 'August 21', isTentative: false },
+  spring: { date: 'January 3', isTentative: false }
+};
+
+const mockUser: IdolMember = {
+  netid: 'abc123',
+  email: 'abc123@cornell.edu',
+  firstName: 'John',
+  lastName: 'Doe',
+  pronouns: '',
+  semesterJoined: '',
+  graduation: 'May 2028',
+  major: 'Information Science',
+  hometown: 'New York, NY',
+  about: `According to all known laws of aviation, there is no way that a bee should be able to fly. Its wings are too small to get its fat little body off the ground. The bee, of course, flies anyways. Because bees don't care what humans think is impossible.\nCrazy? I Was Crazy Once. They Locked Me In A Room. A Rubber Room. A Rubber Room With Rats. And Rats Make Me Crazy.`,
+  subteams: ['idol'],
+  role: 'ops-lead',
+  roleDescription: 'Full Team Lead',
+  doubleMajor: 'Economics',
+  linkedin: 'https://www.linkedin.com/',
+  github: 'https://github.com/',
+  website: 'https://google.com/'
+};
+
 export default function CardPage() {
   return (
     <PageLayout title="Card" description="Guidelines for Card component.">
@@ -70,11 +104,37 @@ export default function CardPage() {
       </PageSection>
 
       <PageSection title="Timeline card" description="Use on the timeline in 'Apply' page">
-        <p className="caps text-accent-green">COMING SOON</p>
+         <TimelineCard event={mockEvent} cycle="freshmen" />
+          <TimelineCard event={{ ...mockEvent, location: undefined }} cycle="freshmen" />
+          <TimelineCard
+            event={{ ...mockEvent, link: undefined, location: 'Upson 102' }}
+            cycle="freshmen"
+          />
       </PageSection>
 
       <PageSection title="Team card" description="Use in the 'Team' page">
-        <p className="caps text-accent-green">COMING SOON</p>
+         <div className="grid lg:grid-cols-4 md:grid-cols-3 sm:grid-cols-2">
+          {Array.from({ length: 6 }).map((_, i) => (
+            <MemberCard
+              key={i}
+              user={{
+                ...mockUser,
+                role: ['ops-lead', 'tpm', 'designer', 'business', 'pm-advisor'][i % 5] as Role,
+                roleDescription: [
+                  'Full Team Lead',
+                  'Technical PM',
+                  'Designer',
+                  'Business',
+                  'PM Advisor'
+                ][i % 5] as RoleDescription
+              }}
+              image={'/clem.jpg'}
+              selected={i > 4}
+            />
+          ))}
+        </div>
+
+        <MemberDetailsCard user={mockUser} image="/clem.jpg" />
       </PageSection>
     </PageLayout>
   );
